refacrot: change library module structure
