+++ src/modules/library/books.controller.ts
@@ -10,7 +10,7 @@ import {
   Query,
   ValidationPipe,
 } from '@nestjs/common';
-import { BookType, LibraryService } from './library.service';
+import { BooksService } from './books.service';
 import { UpdateLibraryDto } from './dto/update-library.dto';
 import { ApiTags } from '@nestjs/swagger';
 import { BooksQueryDto } from './dto/books-query.dto';
@@ -20,11 +20,12 @@ import { CheckAbilities } from 'src/core/decorators/abilities/abilities.decorato
 import { CreatePhysicalBookDto } from './dto/create-physical-book.dto';
 import { CreateDigitalBookDto } from './dto/create-digital-book.dto';
 import { UpdateReferenceDto } from './dto/update-reference.dto';
+import { CreateReferenceDto } from './dto/create-reference.dto';
 
 @ApiTags('Library/Books')
 @Controller('books')
 export class BooksController {
-  constructor(private readonly libraryService: LibraryService) {}
+  constructor(private readonly booksService: BooksService) {}
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'Book' })
@@ -33,7 +34,7 @@ export class BooksController {
     @Body(new ValidationPipe({ transform: true }))
     createDigitalBookDto: CreateDigitalBookDto,
   ) {
-    return this.libraryService.createDigital(createDigitalBookDto);
+    return this.booksService.createDigital(createDigitalBookDto);
   }
 
   @UseGuards(PermissionsGuard)
@@ -43,43 +44,40 @@ export class BooksController {
     @Body(new ValidationPipe({ transform: true }))
     createPhysicalBookDto: CreatePhysicalBookDto,
   ) {
-    return this.libraryService.createPhysical(createPhysicalBookDto);
+    return this.booksService.createPhysical(createPhysicalBookDto);
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'Book' })
   @Get('digital')
   findAllDigital(@Query() queryDto: BooksQueryDto) {
-    return this.libraryService.findAll(queryDto, 'digital');
+    return this.booksService.findAll(queryDto, 'digital');
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'Book' })
   @Get('physical')
   findAllPhysical(@Query() queryDto: BooksQueryDto) {
-    return this.libraryService.findAll(queryDto, 'physical');
+    return this.booksService.findAll(queryDto, 'physical');
   }
 
   @UseGuards(JwtGeneralRequiredAuthGuard)
   @Get('public/digital')
   findAllDigitalPublic(@Query() queryDto: BooksQueryDto) {
-    return this.libraryService.findAll(queryDto, 'digital');
+    return this.booksService.findAll(queryDto, 'digital');
   }
 
   @UseGuards(JwtGeneralRequiredAuthGuard)
   @Get('public/physical')
   findAllPhysicalPublic(@Query() queryDto: BooksQueryDto) {
-    return this.libraryService.findAll(queryDto, 'physical');
+    return this.booksService.findAll(queryDto, 'physical');
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'Book' })
-  @Get('admin/:id/:type')
-  findOneAdmin(
-    @Param('id') id: string,
-    @Param('type') type: 'digital' | 'physical',
-  ) {
-    return this.libraryService.findOne(+id);
+  @Get('admin/:id')
+  findOneAdmin(@Param('id') id: string) {
+    return this.booksService.findOne(+id);
   }
 
   @UseGuards(PermissionsGuard)
@@ -90,50 +88,50 @@ export class BooksController {
     @Body(new ValidationPipe())
     updateLibraryDto: UpdateLibraryDto,
   ) {
-    return this.libraryService.update(+id, updateLibraryDto);
+    return this.booksService.update(+id, updateLibraryDto);
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'LibraryReference' })
-  @Patch('admin/:bookId/:id')
+  @Patch('admin/:bookId/references/:id')
   updateAdminReference(
     @Param('bookId') bookId: string,
     @Param('id') id: string,
     @Body(new ValidationPipe())
     updateReferenceDto: UpdateReferenceDto,
   ) {
-    return this.libraryService.updateReference(+bookId, id, updateReferenceDto);
+    return this.booksService.updateReference(+bookId, id, updateReferenceDto);
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'LibraryReference' })
-  @Post('admin/:bookId/:id')
+  @Post('admin/:bookId/references/:id')
   createLibraryReference(
     @Param('bookId') bookId: string,
     @Param('id') id: string,
     @Body(new ValidationPipe())
-    updateReferenceDto: UpdateReferenceDto,
+    createReferenceDto: CreateReferenceDto,
   ) {
-    return this.libraryService.createReference(+bookId, id, updateReferenceDto);
+    return this.booksService.createReference(+bookId, id, createReferenceDto);
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'LibraryReference' })
-  @Delete('admin/:bookId/:id')
+  @Delete('admin/:bookId/references/:id')
   deleteReference(@Param('bookId') bookId: string, @Param('id') id: string) {
-    return this.libraryService.deleteReference(+bookId, id);
+    return this.booksService.deleteReference(+bookId, id);
   }
 
   @UseGuards(PermissionsGuard)
   @CheckAbilities({ action: 'manage', subject: 'Book' })
   @Delete('admin/:id')
   remove(@Param('id') id: string) {
-    return this.libraryService.remove(+id);
+    return this.booksService.remove(+id);
   }
 
   @UseGuards(JwtGeneralRequiredAuthGuard)
   @Get(':id')
   findOnePublic(@Param('id') id: string) {
-    return this.libraryService.findOnePublic(+id);
+    return this.booksService.findOnePublic(+id);
   }
 }
