+++ src/modules/library/categories.controller.ts
@@ -1,16 +1,16 @@
 import { Controller, Get, UseGuards } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { JwtGeneralRequiredAuthGuard } from 'src/core/guards/jwt-general-required-auth.guard';
-import { LibraryService } from './library.service';
+import { BooksService } from './books.service';
 
 @ApiTags('Library/Categories')
 @Controller('categories')
 export class CategoriesController {
-  constructor(private readonly libraryService: LibraryService) {}
+  constructor(private readonly booksService: BooksService) {}
 
   @UseGuards(JwtGeneralRequiredAuthGuard)
   @Get()
   findAll() {
-    return this.libraryService.findAllCategories();
+    return this.booksService.findAllCategories();
   }
 }
