+++ src/modules/library/dto/create-digital-book.dto.ts
@@ -1,12 +1,4 @@
 import { IntersectionType } from '@nestjs/swagger';
-import { ApiProperty } from '@nestjs/swagger';
-import { Type } from 'class-transformer';
-import {
-  IsNotEmpty,
-  IsString,
-  IsNumber,
-  ValidateNested,
-} from 'class-validator';
 import { CreateLibraryDto } from './create-library.dto';
 import { DigitalBookDto } from './digital-book.dto';
